Refactoring following 8bpp screen driver modification.
- Screen now uses 8bpp resolution
- Colors have been adjusted
- Memory has been optimized to reduce footprint and allow all the tools to be included.

+++ main/img/ble_icons.h
@@ -3,7 +3,7 @@
 #include "twatch.h"
 
 __attribute__ ((aligned(4)))
-extern const uint16_t ble_icons[];
+extern const uint8_t ble_icons[];
 
 #endif /* __INC_BLE_ICONS_H */
