+++ main/img/ble_icons_big.h
@@ -3,7 +3,7 @@
 #include "twatch.h"
 
 __attribute__ ((aligned(4)))
-extern const uint16_t ble_icons_big[];
+extern const uint8_t ble_icons_big[];
 
 #endif /* __INC_BLE_ICONS_BIG_H */
