+++ main/img/ticks.h
@@ -4,6 +4,6 @@
 #include "twatch.h"
 
 __attribute__ ((aligned(4)))
-extern const uint16_t ticks_img[];
+extern const uint8_t ticks_img[];
 
 #endif /* __INC_TICKS_H */
